# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation for NIX library",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://libraryOnLaravel.loc/api",
            "description": "NIX library Laravel Swagger API server"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Outputs the books list",
                "operationId": "examplesAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create new book in library",
                "operationId": "bookCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookShowRequest"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "put": {
                "tags": [
                    "Bookss"
                ],
                "summary": "Update selected book",
                "description": "Returns updated book data",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID for editing",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateBookRequest"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete selected book",
                "description": "Deletes a selected book",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookShowRequest": {
                "title": "Example storring request",
                "description": "The request for book creation",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Book ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "2"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Book title",
                        "type": "string",
                        "example": "Star wars"
                    },
                    "author": {
                        "title": "Author",
                        "description": "Book author",
                        "type": "string",
                        "example": "Pushkin"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Book description",
                        "type": "string",
                        "example": "Short book description"
                    },
                    "book_cover": {
                        "title": "Book cover",
                        "description": "Book cover image",
                        "example": "Some image"
                    },
                    "category": {
                        "title": "Category",
                        "description": "Book category",
                        "type": "string",
                        "format": "string",
                        "example": "history"
                    },
                    "language": {
                        "title": "Language",
                        "description": "Book language",
                        "type": "string",
                        "example": "English"
                    },
                    "publishing_year": {
                        "title": "Publishing year",
                        "description": "The year of book publishing ",
                        "type": "integer",
                        "format": "int32",
                        "example": "1987"
                    },
                    "created_at": {
                        "title": "Book created at",
                        "description": "Book created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-02-02 14:15:03"
                    }
                },
                "type": "object"
            },
            "BookStoreRequest": {
                "title": "Example storring request",
                "description": "The request for book creation",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Book ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Book title",
                        "type": "string",
                        "format": "string",
                        "example": "Star wars"
                    },
                    "author": {
                        "title": "Author",
                        "description": "Book author",
                        "type": "string",
                        "format": "string",
                        "example": "Pushkin"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Book description",
                        "type": "string",
                        "format": "string",
                        "example": "Short book description"
                    },
                    "book_cover": {
                        "title": "Book cover",
                        "description": "Book cover image",
                        "format": "string",
                        "example": "Some image"
                    },
                    "category": {
                        "title": "Category",
                        "description": "Book category",
                        "type": "string",
                        "format": "string",
                        "example": "history"
                    },
                    "language": {
                        "title": "Language",
                        "description": "Book language",
                        "type": "string",
                        "format": "string",
                        "example": "English"
                    },
                    "publishing_year": {
                        "title": "Publishing year",
                        "description": "The year of book publishing ",
                        "type": "integer",
                        "format": "int32",
                        "example": "1987"
                    },
                    "created_at": {
                        "title": "Book created at",
                        "description": "Book created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-02-02 14:15:03"
                    }
                },
                "type": "object"
            },
            "UpdateBookRequest": {
                "title": "Example storring request",
                "description": "The request for book creation",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Book ID",
                        "type": "integer",
                        "format": "int64",
                        "example": "2"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Book title",
                        "type": "string",
                        "example": "Star wars"
                    },
                    "author": {
                        "title": "Author",
                        "description": "Book author",
                        "type": "string",
                        "example": "Pushkin"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Book description",
                        "type": "string",
                        "example": "Short book description"
                    },
                    "book_cover": {
                        "title": "Book cover",
                        "description": "Book cover image",
                        "example": "Some image"
                    },
                    "category": {
                        "title": "Category",
                        "description": "Book category",
                        "type": "string",
                        "format": "string",
                        "example": "history"
                    },
                    "language": {
                        "title": "Language",
                        "description": "Book language",
                        "type": "string",
                        "example": "English"
                    },
                    "publishing_year": {
                        "title": "Publishing year",
                        "description": "The year of book publishing ",
                        "type": "integer",
                        "format": "int32",
                        "example": "1987"
                    },
                    "created_at": {
                        "title": "Book created at",
                        "description": "Book created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-02-02 14:15:03"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "NIX library",
            "description": "Some example pages"
        }
    ]
}